Move insert records to concern
